+++ tests/callbacks/test_callbacks.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
 import pytest
+import torch
 
 import tests.base.utils as tutils
 from pytorch_lightning import Callback
@@ -8,7 +9,6 @@ from pytorch_lightning import Trainer, LightningModule
 from pytorch_lightning.callbacks import EarlyStopping, ModelCheckpoint
 from pytorch_lightning.loggers import TensorBoardLogger
 from tests.base import EvalModelTemplate
-import torch
 
 
 def test_early_stopping_functionality(tmpdir):
