+++ tests/core/test_decorators.py
@@ -1,8 +1,8 @@
 import pytest
 import torch
 
-from tests.base import EvalModelTemplate
 from pytorch_lightning.core.decorators import auto_move_data
+from tests.base import EvalModelTemplate
 
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires GPU machine")
